Deleting event enhancements and Filtered calendar events displayed
Previously when deleting an individual event, the parent record would
never get deleted.  Now the application looks to see if it is the last
event of that parent and if it is, then the parent record is also
deleted.  Also, previously every event was loaded into the calendar for
each call. This would cause a problem when thousands of events were
populated in the table.  Now, only the events that pertain to the
current display are loaded from the databaase