resolved ie redirect error applied audio css
IE required the URL= segment be included in the <meta /> tag that i used
to redirect users from login, changepassword and logout. Without
preceding the URL with that, the browser just entered a loop on the php
page...  firefox didn't have the issue. I also added css that controls
the width and height of my dr. bob audio file becaus in IE it was
appearing too large.