Implemented downloadable pdf of calendar
Implemented html2canvas.js to take a screenshot of the calendar in order
to avert the formatting issues I had with fullcalendar's Print View.
That image gets base64 encoded and sent to a php sub which converts it
to a pdf using mpdf which can then be downloaded and saved to a users
pc.